Creación de 3 interfaces
Se crearon 3 interfaces para poder trabajar de forma grupal
