+++ PROYECTO_ENCOMIENDA/src/proyecto_encomienda/GUI/JFMenu.java
@@ -26,8 +26,6 @@ import proyecto_ecomienda.BDYValidaciones.ConsultarBD;
 import proyecto_ecomienda.BDYValidaciones.CreadorTablas;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import proyecto_ecomienda.BDYValidaciones.ActualizarEstadoFactura;
 import proyecto_ecomienda.BDYValidaciones.Empleados;
 import proyecto_ecomienda.BDYValidaciones.RegistrarDatosFactura;
@@ -43,13 +41,7 @@ import javax.swing.JTable;
 import proyecto_ecomienda.BDYValidaciones.Proveedores;
 import proyecto_ecomienda.BDYValidaciones.ValidadorDeSwings;
 import proyecto_ecomienda.BDYValidaciones.ValidadorDeRegistros;
-import proyecto_encomienda.GESTION_PAQUETES.BACKEND.Inventario;
-import proyecto_encomienda.INCIDENTES.DanioPaquete;
-import proyecto_encomienda.INCIDENTES.ErrorDireccion;
-import proyecto_encomienda.INCIDENTES.GestorIncidentes;
-import proyecto_encomienda.INCIDENTES.Incidente;
-import proyecto_encomienda.INCIDENTES.PaquetePerdido;
-import proyecto_encomienda.INCIDENTES.RechazoEntrega;
+import proyecto_ecomienda.BDYValidaciones.VentanaManager;
 
 /**
  *
@@ -58,17 +50,14 @@ import proyecto_encomienda.INCIDENTES.RechazoEntrega;
 public class JFMenu extends javax.swing.JFrame {
 //Vistas
 
+    private JFInventario inventario = null;
+    private JFRemitente remitente = null;
+    private JFIncidente incidente=null;
+
 //Clases   
     ValidadorDeRegistros validarRegistroF = new ValidadorDeRegistros();
     ValidadorDeSwings validadorCheck = new ValidadorDeSwings();
-//RegistrarDatosFactura facturaRegistrar = new RegistrarDatosFactura();
-    //Clientes
-    private boolean documentoCliente = false;
-    private boolean nombreCliente = false;
-    private boolean apellidoCliente = false;
-    private boolean correoCliente = false;
-    private boolean telefonoCliente = false;
-    private boolean direccionCliente = false;
+
     //Empleados
     private boolean cedulaEmpleadoValidar = false;
     private boolean nombreEmpleadoValidar = false;
@@ -116,28 +105,6 @@ public class JFMenu extends javax.swing.JFrame {
     private boolean dirCliente = false;
     private boolean correoElectronico = false;
     private boolean cambiarSesion = true;
-//Banderas actualizar
-    private boolean telfCliente1 = false;
-    private boolean dirCliente1 = false;
-    private boolean correoElectronico1 = false;
-//Proforma
-    private boolean nombreNegocioValido2 = false;
-    private boolean estadoPagoP2 = false;
-    private boolean rucV2 = false;
-    private boolean dirNeg2 = false;
-    private boolean telefonoNeg2 = false;
-    private boolean rucCliente2 = false;
-    private boolean nombreCFactura2 = false;
-    private boolean apellidoCFactura2 = false;
-    private boolean telfCliente2 = false;
-    private boolean dirCliente2 = false;
-    private boolean correoElectronico2 = false;
-    private boolean cambiarSesion2 = true;
-//Inventario
-    private boolean anchoValidar = false;
-    private boolean largoValidar = false;
-    private boolean pesoValidar = false;
-    private boolean remitenteValidar = false;
 
     private boolean nombreItemValidar1 = false;
     private boolean stockValidar1 = false;
@@ -168,7 +135,6 @@ public class JFMenu extends javax.swing.JFrame {
         String resultado;
         Date fechaYHora = new Date();
         jTCIDelCliente.setEnabled(false);
-        jTFCIRegistrarC.setEnabled(false);
         resultado = mostrarFechaHora(fechaYHora, fecha, localM);
         //jDateChooserFecha.setText(resultado);
         txtID.setText("Usuario     :" + userRol);
@@ -184,7 +150,6 @@ public class JFMenu extends javax.swing.JFrame {
         clickedPanels[4] = Clicked5;
         clickedPanels[5] = Clicked6;
         clickedPanels[6] = Clicked7;
-        contadorIncidentes();
         java.util.Date fechaActual = new java.util.Date();
         // Configura el JDateChooser
         jDateChooserFecha.setMaxSelectableDate(fechaActual); // Fecha máxima permitida
@@ -201,34 +166,13 @@ public class JFMenu extends javax.swing.JFrame {
         //jTFnumerofactura.setText(String.valueOf(nuevoNumeroFactura));
         jTFAdminClave.setToolTipText("Aquí puedes ingresar tu contraseña.");
         desvanecerP();
+        jLInicio.setText("Bienvenido/a");
+    }
+
   
-    }
-
-    public void contadorIncidentes() {
-        try {
-            // Consulta para obtener el máximo ID de Paquete
-            String consulta = "SELECT COALESCE(MAX(IDIncidente), 0) AS max_id FROM Incidente";
-            PreparedStatement stmt = cnx.prepareStatement(consulta);
-            ResultSet rs = stmt.executeQuery();
-
-            if (rs.next()) {
-                int maxId = rs.getInt("max_id");
-                // Si maxId es 0 (no hay paquetes), establece el siguiente ID como 1
-                int siguienteId = (maxId == 0) ? 1 : maxId + 1;
-                IDIncidentesTF.setText(String.valueOf(siguienteId));
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-            // Maneja cualquier error de conexión o consulta aquí
-        }
-    }
-
-    
 
     public void desvanecerP() {
-        JLabel[] labels = {errorActualizarCliente1,
-            errorActualizarCliente2, errorActualizarCliente3,errorActualizarCliente4,
-            errorActualizarCliente5, errorActualizarCliente6, errorProveedores1,
+        JLabel[] labels = {errorProveedores1,
             errorProveedores2, errorProveedores3, errorProveedores4, errorProveedores5, errorProveedores6, errorProveedores7,
             errorProveedores8,
             jLErrorEmpleado1, jLErrorEmpleado2, jLErrorEmpleado3,
@@ -261,12 +205,6 @@ public class JFMenu extends javax.swing.JFrame {
         jError5.setVisible(false);
         jError6.setVisible(false);
         jCBEstadoPago.setEnabled(false);
-        errorc1.setVisible(false);
-        errorc2.setVisible(false);
-        errorc3.setVisible(false);
-        errorc4.setVisible(false);
-        errorc5.setVisible(false);
-        errorc6.setVisible(false);
         jError7.setVisible(false);
         jError8.setVisible(false);
         jError9.setVisible(false);
@@ -323,69 +261,9 @@ public class JFMenu extends javax.swing.JFrame {
         btnHelp = new javax.swing.JButton();
         panelContent = new javax.swing.JPanel();
         JPGPaquetes = new javax.swing.JPanel();
-        jLabel6 = new javax.swing.JLabel();
+        jLInicio = new javax.swing.JLabel();
         JPRemitente = new javax.swing.JPanel();
         jLabel5 = new javax.swing.JLabel();
-        jPClientes = new javax.swing.JTabbedPane();
-        jPPR1 = new javax.swing.JPanel();
-        jScrollPane16 = new javax.swing.JScrollPane();
-        jTablaClientesR = new javax.swing.JTable();
-        jPanel46 = new javax.swing.JPanel();
-        jLabel66 = new javax.swing.JLabel();
-        jTFNombresR = new javax.swing.JTextField();
-        jLabel71 = new javax.swing.JLabel();
-        jTFTelefonoR = new javax.swing.JTextField();
-        jLTipoCli = new javax.swing.JLabel();
-        jTFCIRegistrarC = new javax.swing.JTextField();
-        jLabel73 = new javax.swing.JLabel();
-        jTFDireccionR = new javax.swing.JTextField();
-        jTFApellidosR = new javax.swing.JTextField();
-        jLabel89 = new javax.swing.JLabel();
-        correoCli2 = new javax.swing.JTextField();
-        jLabel87 = new javax.swing.JLabel();
-        errorc1 = new javax.swing.JLabel();
-        errorc2 = new javax.swing.JLabel();
-        errorc3 = new javax.swing.JLabel();
-        errorc4 = new javax.swing.JLabel();
-        errorc5 = new javax.swing.JLabel();
-        errorc6 = new javax.swing.JLabel();
-        jPanel37 = new javax.swing.JPanel();
-        jCBNacionalExtranjero = new javax.swing.JComboBox<>();
-        jCJuridicoNatural = new javax.swing.JComboBox<>();
-        jBRegistarCliente = new javax.swing.JButton();
-        jPPCE1 = new javax.swing.JPanel();
-        jPCambiarEstado = new javax.swing.JPanel();
-        jPPA1 = new javax.swing.JPanel();
-        jPActualizarClientes = new javax.swing.JPanel();
-        jPanel40 = new javax.swing.JPanel();
-        jBIActualizarAct1 = new javax.swing.JButton();
-        jPanel43 = new javax.swing.JPanel();
-        jTFIBuscadorAct1 = new javax.swing.JTextField();
-        documentoActualizar = new javax.swing.JLabel();
-        CIselect = new javax.swing.JCheckBox();
-        nombreSelect = new javax.swing.JCheckBox();
-        apellidoSelect = new javax.swing.JCheckBox();
-        apellidoActualizar = new javax.swing.JTextField();
-        nombreActualizar = new javax.swing.JTextField();
-        ciActualizar = new javax.swing.JTextField();
-        telefonoSelect = new javax.swing.JCheckBox();
-        direccionSelect = new javax.swing.JCheckBox();
-        direccionActualizar = new javax.swing.JTextField();
-        telefonoActualizar = new javax.swing.JTextField();
-        correoSelect = new javax.swing.JCheckBox();
-        correoActualizar = new javax.swing.JTextField();
-        errorActualizarCliente1 = new javax.swing.JLabel();
-        errorActualizarCliente2 = new javax.swing.JLabel();
-        errorActualizarCliente3 = new javax.swing.JLabel();
-        errorActualizarCliente4 = new javax.swing.JLabel();
-        errorActualizarCliente5 = new javax.swing.JLabel();
-        errorActualizarCliente6 = new javax.swing.JLabel();
-        jScrollPane18 = new javax.swing.JScrollPane();
-        tablaActualizarCliente = new javax.swing.JTable();
-        jPanel38 = new javax.swing.JPanel();
-        jCBNacionalExtranjero2 = new javax.swing.JComboBox<>();
-        jCJuridicoNatural1 = new javax.swing.JComboBox<>();
-        jPPC1 = new javax.swing.JPanel();
         JPConductores = new javax.swing.JPanel();
         jLabel7 = new javax.swing.JLabel();
         jPGP = new javax.swing.JTabbedPane();
@@ -627,25 +505,6 @@ public class JFMenu extends javax.swing.JFrame {
         jLabel26 = new javax.swing.JLabel();
         JPIncidentes = new javax.swing.JPanel();
         jLabel117 = new javax.swing.JLabel();
-        jPIncidentes = new javax.swing.JPanel();
-        jTPEmpleados1 = new javax.swing.JTabbedPane();
-        jPRE1 = new javax.swing.JPanel();
-        jPanel10 = new javax.swing.JPanel();
-        SeleccionIncidentesCB = new javax.swing.JComboBox<>();
-        jLabel133 = new javax.swing.JLabel();
-        jLabel134 = new javax.swing.JLabel();
-        IDPIncidentesTF = new javax.swing.JTextField();
-        jLabel135 = new javax.swing.JLabel();
-        jScrollPane4 = new javax.swing.JScrollPane();
-        RazonIncidentesTA = new javax.swing.JTextArea();
-        RegistrarIncidentesButton = new javax.swing.JButton();
-        jLabel136 = new javax.swing.JLabel();
-        IDIncidentesTF = new javax.swing.JTextField();
-        jScrollPane28 = new javax.swing.JScrollPane();
-        jTIncidentes = new javax.swing.JTable();
-        jPAE1 = new javax.swing.JPanel();
-        jPEE1 = new javax.swing.JPanel();
-        jLabel132 = new javax.swing.JLabel();
         JPAdministración = new javax.swing.JPanel();
         jLabel25 = new javax.swing.JLabel();
         jTPAdmin = new javax.swing.JTabbedPane();
@@ -708,6 +567,7 @@ public class JFMenu extends javax.swing.JFrame {
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("SistemaBestColor");
         setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        setUndecorated(true);
         setResizable(false);
 
         Home.setLayout(new java.awt.CardLayout());
@@ -1076,634 +936,20 @@ public class JFMenu extends javax.swing.JFrame {
         JPGPaquetes.setBackground(new java.awt.Color(245, 245, 245));
         JPGPaquetes.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
 
-        jLabel6.setFont(new java.awt.Font("Tahoma", 0, 64)); // NOI18N
-        jLabel6.setForeground(new java.awt.Color(102, 102, 102));
-        jLabel6.setText("Gestión de Paquetes");
-        JPGPaquetes.add(jLabel6, new org.netbeans.lib.awtextra.AbsoluteConstraints(300, 220, 620, 250));
+        jLInicio.setFont(new java.awt.Font("Tahoma", 0, 64)); // NOI18N
+        jLInicio.setForeground(new java.awt.Color(102, 102, 102));
+        jLInicio.setText("Gestión de Paquetes");
+        JPGPaquetes.add(jLInicio, new org.netbeans.lib.awtextra.AbsoluteConstraints(300, 220, 620, 250));
 
         panelContent.add(JPGPaquetes, "card1");
 
         JPRemitente.setBackground(new java.awt.Color(245, 245, 245));
         JPRemitente.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
 
-        jLabel5.setFont(new java.awt.Font("Tahoma", 0, 24)); // NOI18N
+        jLabel5.setFont(new java.awt.Font("Tahoma", 0, 62)); // NOI18N
         jLabel5.setForeground(new java.awt.Color(102, 102, 102));
         jLabel5.setText("Remitente");
-        JPRemitente.add(jLabel5, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 10, -1, -1));
-
-        jPPR1.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        jTablaClientesR.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        jScrollPane16.setViewportView(jTablaClientesR);
-
-        jPPR1.add(jScrollPane16, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 380, 1000, 158));
-
-        jPanel46.setBorder(javax.swing.BorderFactory.createTitledBorder("Datos Cliente"));
-
-        jLabel66.setText("Nombres");
-
-        jTFNombresR.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                jTFNombresRFocusLost(evt);
-            }
-        });
-        jTFNombresR.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTFNombresRKeyReleased(evt);
-            }
-            public void keyTyped(java.awt.event.KeyEvent evt) {
-                jTFNombresRKeyTyped(evt);
-            }
-        });
-
-        jLabel71.setText("Teléfono");
-
-        jTFTelefonoR.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                jTFTelefonoRFocusLost(evt);
-            }
-        });
-        jTFTelefonoR.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTFTelefonoRKeyReleased(evt);
-            }
-        });
-
-        jLTipoCli.setText("CI");
-
-        jTFCIRegistrarC.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                jTFCIRegistrarCFocusLost(evt);
-            }
-        });
-        jTFCIRegistrarC.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTFCIRegistrarCKeyReleased(evt);
-            }
-        });
-
-        jLabel73.setText("Dirección");
-
-        jTFDireccionR.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                jTFDireccionRFocusLost(evt);
-            }
-        });
-        jTFDireccionR.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTFDireccionRKeyReleased(evt);
-            }
-        });
-
-        jTFApellidosR.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                jTFApellidosRFocusLost(evt);
-            }
-        });
-        jTFApellidosR.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTFApellidosRKeyReleased(evt);
-            }
-            public void keyTyped(java.awt.event.KeyEvent evt) {
-                jTFApellidosRKeyTyped(evt);
-            }
-        });
-
-        jLabel89.setText("Apellidos");
-
-        correoCli2.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                correoCli2FocusLost(evt);
-            }
-        });
-        correoCli2.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                correoCli2KeyReleased(evt);
-            }
-        });
-
-        jLabel87.setText("Correo");
-
-        errorc1.setForeground(new java.awt.Color(255, 102, 102));
-        errorc1.setText("*error");
-
-        errorc2.setForeground(new java.awt.Color(255, 102, 102));
-        errorc2.setText("Nombre inválido");
-
-        errorc3.setForeground(new java.awt.Color(255, 102, 102));
-        errorc3.setText("Apellido inválido");
-
-        errorc4.setForeground(new java.awt.Color(255, 102, 102));
-        errorc4.setText("Correo inválido");
-
-        errorc5.setForeground(new java.awt.Color(255, 102, 102));
-        errorc5.setText("*error");
-
-        errorc6.setForeground(new java.awt.Color(255, 102, 102));
-        errorc6.setText("Dirección inválida");
-
-        javax.swing.GroupLayout jPanel46Layout = new javax.swing.GroupLayout(jPanel46);
-        jPanel46.setLayout(jPanel46Layout);
-        jPanel46Layout.setHorizontalGroup(
-            jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel46Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel46Layout.createSequentialGroup()
-                        .addComponent(jLabel89)
-                        .addGap(18, 18, 18)
-                        .addComponent(errorc3)
-                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .addGroup(jPanel46Layout.createSequentialGroup()
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                            .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel46Layout.createSequentialGroup()
-                                .addComponent(jLabel87, javax.swing.GroupLayout.PREFERRED_SIZE, 43, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(18, 18, 18)
-                                .addComponent(errorc4))
-                            .addComponent(jTFApellidosR, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.DEFAULT_SIZE, 182, Short.MAX_VALUE)
-                            .addComponent(correoCli2, javax.swing.GroupLayout.Alignment.LEADING))
-                        .addGap(0, 0, Short.MAX_VALUE))
-                    .addGroup(jPanel46Layout.createSequentialGroup()
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                            .addComponent(jTFCIRegistrarC, javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                .addGroup(jPanel46Layout.createSequentialGroup()
-                                    .addComponent(jLabel66)
-                                    .addGap(18, 18, 18)
-                                    .addComponent(errorc2))
-                                .addComponent(jTFNombresR, javax.swing.GroupLayout.PREFERRED_SIZE, 182, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGroup(jPanel46Layout.createSequentialGroup()
-                                    .addComponent(jLTipoCli)
-                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                    .addComponent(errorc1))))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 70, Short.MAX_VALUE)
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jTFDireccionR, javax.swing.GroupLayout.PREFERRED_SIZE, 165, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addGroup(jPanel46Layout.createSequentialGroup()
-                                .addComponent(jLabel73)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                .addComponent(errorc6))
-                            .addComponent(jTFTelefonoR, javax.swing.GroupLayout.PREFERRED_SIZE, 165, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addGroup(jPanel46Layout.createSequentialGroup()
-                                .addComponent(jLabel71)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(errorc5)))
-                        .addGap(45, 45, 45))))
-        );
-        jPanel46Layout.setVerticalGroup(
-            jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel46Layout.createSequentialGroup()
-                .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel46Layout.createSequentialGroup()
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLTipoCli)
-                            .addComponent(errorc1))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addComponent(jTFCIRegistrarC, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel66)
-                            .addComponent(errorc2))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTFNombresR, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPanel46Layout.createSequentialGroup()
-                        .addContainerGap()
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel71)
-                            .addComponent(errorc5))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTFTelefonoR, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel73)
-                            .addComponent(errorc6))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTFDireccionR, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel89)
-                    .addComponent(errorc3))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jTFApellidosR, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel46Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel87)
-                    .addComponent(errorc4))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(correoCli2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(14, 14, 14))
-        );
-
-        jPPR1.add(jPanel46, new org.netbeans.lib.awtextra.AbsoluteConstraints(237, 23, -1, 241));
-
-        jPanel37.setBorder(javax.swing.BorderFactory.createTitledBorder("Selecciona el tipo de cliente"));
-
-        jCBNacionalExtranjero.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Selecciona", "Nacional", "Extranjero" }));
-        jCBNacionalExtranjero.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jCBNacionalExtranjeroActionPerformed(evt);
-            }
-        });
-
-        jCJuridicoNatural.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Natural o Jurídico", "Natural", "Jurídico" }));
-        jCJuridicoNatural.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jCJuridicoNaturalActionPerformed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jPanel37Layout = new javax.swing.GroupLayout(jPanel37);
-        jPanel37.setLayout(jPanel37Layout);
-        jPanel37Layout.setHorizontalGroup(
-            jPanel37Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel37Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel37Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jCBNacionalExtranjero, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jCJuridicoNatural, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(61, Short.MAX_VALUE))
-        );
-        jPanel37Layout.setVerticalGroup(
-            jPanel37Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel37Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jCBNacionalExtranjero, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(jCJuridicoNatural, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(24, Short.MAX_VALUE))
-        );
-
-        jPPR1.add(jPanel37, new org.netbeans.lib.awtextra.AbsoluteConstraints(26, 23, -1, -1));
-
-        jBRegistarCliente.setText("Registrar Cliente");
-        jBRegistarCliente.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBRegistarClienteActionPerformed(evt);
-            }
-        });
-        jPPR1.add(jBRegistarCliente, new org.netbeans.lib.awtextra.AbsoluteConstraints(380, 310, -1, -1));
-
-        jPClientes.addTab("Registrar", jPPR1);
-
-        jPPCE1.setLayout(new java.awt.CardLayout());
-
-        javax.swing.GroupLayout jPCambiarEstadoLayout = new javax.swing.GroupLayout(jPCambiarEstado);
-        jPCambiarEstado.setLayout(jPCambiarEstadoLayout);
-        jPCambiarEstadoLayout.setHorizontalGroup(
-            jPCambiarEstadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1060, Short.MAX_VALUE)
-        );
-        jPCambiarEstadoLayout.setVerticalGroup(
-            jPCambiarEstadoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 575, Short.MAX_VALUE)
-        );
-
-        jPPCE1.add(jPCambiarEstado, "card1");
-
-        jPClientes.addTab("Cambiar Estado", jPPCE1);
-
-        jPPA1.setPreferredSize(new java.awt.Dimension(790, 459));
-        jPPA1.setLayout(new java.awt.CardLayout());
-
-        jPanel40.setBorder(javax.swing.BorderFactory.createTitledBorder("Actualizar Cliente"));
-        jPanel40.setToolTipText("");
-        jPanel40.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-
-        jBIActualizarAct1.setText("Actualizar");
-        jBIActualizarAct1.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBIActualizarAct1ActionPerformed(evt);
-            }
-        });
-
-        jPanel43.setBorder(javax.swing.BorderFactory.createTitledBorder(null, "Buscardor", javax.swing.border.TitledBorder.LEFT, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Times New Roman", 0, 12))); // NOI18N
-
-        jTFIBuscadorAct1.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTFIBuscadorAct1KeyReleased(evt);
-            }
-            public void keyTyped(java.awt.event.KeyEvent evt) {
-                jTFIBuscadorAct1KeyTyped(evt);
-            }
-        });
-
-        documentoActualizar.setText("Documento");
-
-        javax.swing.GroupLayout jPanel43Layout = new javax.swing.GroupLayout(jPanel43);
-        jPanel43.setLayout(jPanel43Layout);
-        jPanel43Layout.setHorizontalGroup(
-            jPanel43Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel43Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(documentoActualizar)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 51, Short.MAX_VALUE)
-                .addComponent(jTFIBuscadorAct1, javax.swing.GroupLayout.PREFERRED_SIZE, 171, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(28, 28, 28))
-        );
-        jPanel43Layout.setVerticalGroup(
-            jPanel43Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel43Layout.createSequentialGroup()
-                .addGroup(jPanel43Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jTFIBuscadorAct1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(documentoActualizar))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        CIselect.setSelected(true);
-        CIselect.setText("CI");
-        CIselect.setEnabled(false);
-
-        nombreSelect.setSelected(true);
-        nombreSelect.setText("Nombre");
-        nombreSelect.setEnabled(false);
-
-        apellidoSelect.setSelected(true);
-        apellidoSelect.setText("Apellido");
-        apellidoSelect.setEnabled(false);
-
-        apellidoActualizar.setEditable(false);
-
-        nombreActualizar.setEditable(false);
-
-        ciActualizar.setEditable(false);
-
-        telefonoSelect.setText("Teléfono");
-        telefonoSelect.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                telefonoSelectActionPerformed(evt);
-            }
-        });
-
-        direccionSelect.setText("Dirección");
-        direccionSelect.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                direccionSelectActionPerformed(evt);
-            }
-        });
-
-        direccionActualizar.setEnabled(false);
-        direccionActualizar.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                direccionActualizarFocusLost(evt);
-            }
-        });
-        direccionActualizar.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                direccionActualizarKeyReleased(evt);
-            }
-        });
-
-        telefonoActualizar.setEnabled(false);
-        telefonoActualizar.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                telefonoActualizarFocusLost(evt);
-            }
-        });
-        telefonoActualizar.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                telefonoActualizarKeyReleased(evt);
-            }
-        });
-
-        correoSelect.setText("Correo");
-        correoSelect.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                correoSelectActionPerformed(evt);
-            }
-        });
-
-        correoActualizar.setEnabled(false);
-        correoActualizar.addFocusListener(new java.awt.event.FocusAdapter() {
-            public void focusLost(java.awt.event.FocusEvent evt) {
-                correoActualizarFocusLost(evt);
-            }
-        });
-        correoActualizar.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                correoActualizarKeyReleased(evt);
-            }
-        });
-
-        errorActualizarCliente1.setForeground(new java.awt.Color(255, 0, 51));
-        errorActualizarCliente1.setText("*error");
-
-        errorActualizarCliente2.setForeground(new java.awt.Color(255, 0, 51));
-        errorActualizarCliente2.setText("*Dirección inválida");
-
-        errorActualizarCliente3.setForeground(new java.awt.Color(255, 0, 51));
-        errorActualizarCliente3.setText("Correo inválido");
-
-        errorActualizarCliente4.setForeground(new java.awt.Color(255, 0, 51));
-        errorActualizarCliente4.setText("*Vacio");
-
-        errorActualizarCliente5.setForeground(new java.awt.Color(255, 0, 51));
-        errorActualizarCliente5.setText("*Vacio");
-
-        errorActualizarCliente6.setForeground(new java.awt.Color(255, 0, 51));
-        errorActualizarCliente6.setText("*Vacio");
-
-        javax.swing.GroupLayout jPanel40Layout = new javax.swing.GroupLayout(jPanel40);
-        jPanel40.setLayout(jPanel40Layout);
-        jPanel40Layout.setHorizontalGroup(
-            jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel40Layout.createSequentialGroup()
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel40Layout.createSequentialGroup()
-                        .addContainerGap()
-                        .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jPanel43, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(telefonoSelect, javax.swing.GroupLayout.PREFERRED_SIZE, 91, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addGroup(jPanel40Layout.createSequentialGroup()
-                                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addGroup(jPanel40Layout.createSequentialGroup()
-                                        .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                            .addComponent(nombreSelect)
-                                            .addComponent(CIselect))
-                                        .addGap(18, 18, 18)
-                                        .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                            .addComponent(correoActualizar, javax.swing.GroupLayout.DEFAULT_SIZE, 196, Short.MAX_VALUE)
-                                            .addComponent(ciActualizar)
-                                            .addComponent(direccionActualizar)
-                                            .addComponent(telefonoActualizar)))
-                                    .addComponent(correoSelect, javax.swing.GroupLayout.PREFERRED_SIZE, 91, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                                        .addComponent(nombreActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, 196, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                        .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                            .addComponent(direccionSelect, javax.swing.GroupLayout.PREFERRED_SIZE, 91, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                            .addGroup(jPanel40Layout.createSequentialGroup()
-                                                .addComponent(apellidoSelect)
-                                                .addGap(18, 18, 18)
-                                                .addComponent(apellidoActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, 194, javax.swing.GroupLayout.PREFERRED_SIZE)))))
-                                .addGap(18, 18, 18)
-                                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(errorActualizarCliente2)
-                                    .addComponent(errorActualizarCliente3)
-                                    .addComponent(errorActualizarCliente1)
-                                    .addComponent(errorActualizarCliente4)
-                                    .addComponent(errorActualizarCliente5)
-                                    .addComponent(errorActualizarCliente6)))))
-                    .addGroup(jPanel40Layout.createSequentialGroup()
-                        .addGap(153, 153, 153)
-                        .addComponent(jBIActualizarAct1)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        jPanel40Layout.setVerticalGroup(
-            jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel40Layout.createSequentialGroup()
-                .addGap(18, 18, 18)
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(CIselect)
-                    .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(ciActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(errorActualizarCliente4)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(nombreActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(nombreSelect)
-                    .addComponent(errorActualizarCliente5))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(apellidoActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(apellidoSelect)
-                    .addComponent(errorActualizarCliente6))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(telefonoSelect)
-                    .addComponent(telefonoActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(errorActualizarCliente1))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(direccionSelect)
-                    .addComponent(direccionActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(errorActualizarCliente2))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGroup(jPanel40Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(correoSelect)
-                    .addComponent(correoActualizar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(errorActualizarCliente3))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jPanel43, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jBIActualizarAct1)
-                .addGap(14, 14, 14))
-        );
-
-        tablaActualizarCliente.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        tablaActualizarCliente.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                tablaActualizarClienteMouseClicked(evt);
-            }
-        });
-        jScrollPane18.setViewportView(tablaActualizarCliente);
-
-        jPanel38.setBorder(javax.swing.BorderFactory.createTitledBorder("Selecciona el tipo de cliente"));
-
-        jCBNacionalExtranjero2.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Selecciona", "Nacional", "Extranjero" }));
-        jCBNacionalExtranjero2.setEnabled(false);
-        jCBNacionalExtranjero2.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jCBNacionalExtranjero2ActionPerformed(evt);
-            }
-        });
-
-        jCJuridicoNatural1.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Natural o Jurídico", "Natural", "Jurídico" }));
-        jCJuridicoNatural1.setEnabled(false);
-        jCJuridicoNatural1.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jCJuridicoNatural1ActionPerformed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jPanel38Layout = new javax.swing.GroupLayout(jPanel38);
-        jPanel38.setLayout(jPanel38Layout);
-        jPanel38Layout.setHorizontalGroup(
-            jPanel38Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel38Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel38Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jCBNacionalExtranjero2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jCJuridicoNatural1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(61, Short.MAX_VALUE))
-        );
-        jPanel38Layout.setVerticalGroup(
-            jPanel38Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel38Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jCBNacionalExtranjero2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(jCJuridicoNatural1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(24, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout jPActualizarClientesLayout = new javax.swing.GroupLayout(jPActualizarClientes);
-        jPActualizarClientes.setLayout(jPActualizarClientesLayout);
-        jPActualizarClientesLayout.setHorizontalGroup(
-            jPActualizarClientesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPActualizarClientesLayout.createSequentialGroup()
-                .addGroup(jPActualizarClientesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPActualizarClientesLayout.createSequentialGroup()
-                        .addGap(56, 56, 56)
-                        .addComponent(jPanel38, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(50, 50, 50)
-                        .addComponent(jPanel40, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPActualizarClientesLayout.createSequentialGroup()
-                        .addGap(32, 32, 32)
-                        .addComponent(jScrollPane18, javax.swing.GroupLayout.PREFERRED_SIZE, 958, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(70, Short.MAX_VALUE))
-        );
-        jPActualizarClientesLayout.setVerticalGroup(
-            jPActualizarClientesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPActualizarClientesLayout.createSequentialGroup()
-                .addGroup(jPActualizarClientesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPActualizarClientesLayout.createSequentialGroup()
-                        .addContainerGap(9, Short.MAX_VALUE)
-                        .addComponent(jPanel40, javax.swing.GroupLayout.PREFERRED_SIZE, 304, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPActualizarClientesLayout.createSequentialGroup()
-                        .addGap(44, 44, 44)
-                        .addComponent(jPanel38, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addGap(18, 18, 18)
-                .addComponent(jScrollPane18, javax.swing.GroupLayout.PREFERRED_SIZE, 223, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(21, 21, 21))
-        );
-
-        jPPA1.add(jPActualizarClientes, "card1");
-
-        jPClientes.addTab("Actualizar", jPPA1);
-
-        javax.swing.GroupLayout jPPC1Layout = new javax.swing.GroupLayout(jPPC1);
-        jPPC1.setLayout(jPPC1Layout);
-        jPPC1Layout.setHorizontalGroup(
-            jPPC1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1060, Short.MAX_VALUE)
-        );
-        jPPC1Layout.setVerticalGroup(
-            jPPC1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 575, Short.MAX_VALUE)
-        );
-
-        jPClientes.addTab("Consultar", jPPC1);
-
-        JPRemitente.add(jPClientes, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 60, 1060, 610));
+        JPRemitente.add(jLabel5, new org.netbeans.lib.awtextra.AbsoluteConstraints(410, 310, -1, -1));
 
         panelContent.add(JPRemitente, "card2");
 
@@ -4057,192 +3303,10 @@ public class JFMenu extends javax.swing.JFrame {
         JPIncidentes.setBackground(new java.awt.Color(245, 245, 245));
         JPIncidentes.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
 
-        jLabel117.setFont(new java.awt.Font("Tahoma", 0, 24)); // NOI18N
+        jLabel117.setFont(new java.awt.Font("Tahoma", 0, 64)); // NOI18N
         jLabel117.setForeground(new java.awt.Color(102, 102, 102));
         jLabel117.setText("Incidentes");
-        JPIncidentes.add(jLabel117, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 10, -1, -1));
-
-        jTPEmpleados1.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jTPEmpleados1MouseClicked(evt);
-            }
-        });
-
-        jPanel10.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
-
-        SeleccionIncidentesCB.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Selecciona", "Daño en el Paquete", "Error de Dirección", "Paquete Perdido", "Rechazo Entrega" }));
-        SeleccionIncidentesCB.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                SeleccionIncidentesCBActionPerformed(evt);
-            }
-        });
-
-        jLabel133.setText("Tipo de incidente:");
-
-        jLabel134.setText("Código único del paquete");
-
-        IDPIncidentesTF.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                IDPIncidentesTFKeyReleased(evt);
-            }
-        });
-
-        jLabel135.setText("Descripción");
-
-        RazonIncidentesTA.setColumns(20);
-        RazonIncidentesTA.setRows(5);
-        jScrollPane4.setViewportView(RazonIncidentesTA);
-
-        RegistrarIncidentesButton.setText("Registrar");
-        RegistrarIncidentesButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                RegistrarIncidentesButtonActionPerformed(evt);
-            }
-        });
-
-        jLabel136.setText("Código único del incidente");
-
-        IDIncidentesTF.setEditable(false);
-        IDIncidentesTF.setEnabled(false);
-        IDIncidentesTF.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                IDIncidentesTFActionPerformed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jPanel10Layout = new javax.swing.GroupLayout(jPanel10);
-        jPanel10.setLayout(jPanel10Layout);
-        jPanel10Layout.setHorizontalGroup(
-            jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel10Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                        .addComponent(jLabel134, javax.swing.GroupLayout.DEFAULT_SIZE, 163, Short.MAX_VALUE)
-                        .addComponent(jLabel133, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .addComponent(jLabel135)
-                    .addComponent(jLabel136, javax.swing.GroupLayout.PREFERRED_SIZE, 181, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 118, Short.MAX_VALUE)
-                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(RegistrarIncidentesButton)
-                    .addComponent(jScrollPane4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                        .addComponent(SeleccionIncidentesCB, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(IDPIncidentesTF, javax.swing.GroupLayout.PREFERRED_SIZE, 138, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addComponent(IDIncidentesTF, javax.swing.GroupLayout.PREFERRED_SIZE, 138, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(124, Short.MAX_VALUE))
-        );
-        jPanel10Layout.setVerticalGroup(
-            jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel10Layout.createSequentialGroup()
-                .addGap(15, 15, 15)
-                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel136)
-                    .addComponent(IDIncidentesTF, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel134)
-                    .addComponent(IDPIncidentesTF, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel133)
-                    .addComponent(SeleccionIncidentesCB, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(20, 20, 20)
-                .addGroup(jPanel10Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jScrollPane4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel135))
-                .addGap(18, 18, 18)
-                .addComponent(RegistrarIncidentesButton)
-                .addContainerGap(74, Short.MAX_VALUE))
-        );
-
-        jTIncidentes.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null},
-                {null, null, null, null}
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        jScrollPane28.setViewportView(jTIncidentes);
-
-        javax.swing.GroupLayout jPRE1Layout = new javax.swing.GroupLayout(jPRE1);
-        jPRE1.setLayout(jPRE1Layout);
-        jPRE1Layout.setHorizontalGroup(
-            jPRE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPRE1Layout.createSequentialGroup()
-                .addGroup(jPRE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPRE1Layout.createSequentialGroup()
-                        .addGap(38, 38, 38)
-                        .addComponent(jPanel10, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPRE1Layout.createSequentialGroup()
-                        .addGap(15, 15, 15)
-                        .addComponent(jScrollPane28, javax.swing.GroupLayout.PREFERRED_SIZE, 734, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(216, Short.MAX_VALUE))
-        );
-        jPRE1Layout.setVerticalGroup(
-            jPRE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPRE1Layout.createSequentialGroup()
-                .addGap(24, 24, 24)
-                .addComponent(jPanel10, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(jScrollPane28, javax.swing.GroupLayout.PREFERRED_SIZE, 153, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(31, Short.MAX_VALUE))
-        );
-
-        jTPEmpleados1.addTab("Registrar Incidente", jPRE1);
-
-        javax.swing.GroupLayout jPAE1Layout = new javax.swing.GroupLayout(jPAE1);
-        jPAE1.setLayout(jPAE1Layout);
-        jPAE1Layout.setHorizontalGroup(
-            jPAE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 965, Short.MAX_VALUE)
-        );
-        jPAE1Layout.setVerticalGroup(
-            jPAE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 566, Short.MAX_VALUE)
-        );
-
-        jTPEmpleados1.addTab("Actualizar", jPAE1);
-
-        javax.swing.GroupLayout jPEE1Layout = new javax.swing.GroupLayout(jPEE1);
-        jPEE1.setLayout(jPEE1Layout);
-        jPEE1Layout.setHorizontalGroup(
-            jPEE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 965, Short.MAX_VALUE)
-        );
-        jPEE1Layout.setVerticalGroup(
-            jPEE1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 566, Short.MAX_VALUE)
-        );
-
-        jTPEmpleados1.addTab("Cambiar de estado", jPEE1);
-
-        javax.swing.GroupLayout jPIncidentesLayout = new javax.swing.GroupLayout(jPIncidentes);
-        jPIncidentes.setLayout(jPIncidentesLayout);
-        jPIncidentesLayout.setHorizontalGroup(
-            jPIncidentesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPIncidentesLayout.createSequentialGroup()
-                .addGap(36, 36, 36)
-                .addComponent(jTPEmpleados1, javax.swing.GroupLayout.PREFERRED_SIZE, 965, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(79, Short.MAX_VALUE))
-        );
-        jPIncidentesLayout.setVerticalGroup(
-            jPIncidentesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPIncidentesLayout.createSequentialGroup()
-                .addGap(0, 39, Short.MAX_VALUE)
-                .addComponent(jTPEmpleados1, javax.swing.GroupLayout.PREFERRED_SIZE, 601, javax.swing.GroupLayout.PREFERRED_SIZE))
-        );
-
-        JPIncidentes.add(jPIncidentes, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 40, 1080, 640));
-
-        jLabel132.setFont(new java.awt.Font("Tahoma", 0, 48)); // NOI18N
-        jLabel132.setForeground(new java.awt.Color(102, 102, 102));
-        jLabel132.setText("Solo Administradores");
-        JPIncidentes.add(jLabel132, new org.netbeans.lib.awtextra.AbsoluteConstraints(200, 260, -1, -1));
+        JPIncidentes.add(jLabel117, new org.netbeans.lib.awtextra.AbsoluteConstraints(360, 270, -1, -1));
 
         panelContent.add(JPIncidentes, "card6");
 
@@ -4950,11 +4014,12 @@ public class JFMenu extends javax.swing.JFrame {
         contenido.show(panelContent, "card1");
         cambiarSeccionMenu(0);
         menuinventario.setBackground(Color.decode("#494848"));
-        JFInventario inventario=new JFInventario(cnx);
-        inventario.setVisible(true);
+        JFrame ventanaInventario = new JFInventario(cnx);
+        VentanaManager.getInstance().mostrarVentana("inventario", ventanaInventario);
+        jLInicio.setText("Paquetes");
     }//GEN-LAST:event_menuinventarioMouseClicked
 
-    
+
     private void menuProveedoresMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_menuProveedoresMouseClicked
         contenido.show(panelContent, "card3");
         cambiarSeccionMenu(2);
@@ -4994,6 +4059,8 @@ public class JFMenu extends javax.swing.JFrame {
         contenido.show(panelContent, "card2");
         cambiarSeccionMenu(1);
         menuClientes.setBackground(Color.decode("#494848"));
+        JFrame remintente = new JFRemitente(cnx);
+        VentanaManager.getInstance().mostrarVentana("remintente", remintente);
     }//GEN-LAST:event_menuClientesMouseClicked
 
     private void menuClientesMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_menuClientesMouseEntered
@@ -5185,169 +4252,6 @@ public class JFMenu extends javax.swing.JFrame {
     private void jTFIBuscarRUCKeyTyped(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFIBuscarRUCKeyTyped
         // TODO add your handling code here:
     }//GEN-LAST:event_jTFIBuscarRUCKeyTyped
-
-    private void jTFIBuscadorAct1KeyTyped(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFIBuscadorAct1KeyTyped
-        // TODO add your handling code here:
-    }//GEN-LAST:event_jTFIBuscadorAct1KeyTyped
-
-    private void jTFIBuscadorAct1KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFIBuscadorAct1KeyReleased
-        String filtroIdCliente = jTFIBuscadorAct1.getText();
-        CreadorTablas creadorTablas = new CreadorTablas();
-        DefaultTableModel modelo = creadorTablas.mostrarClienteParametrizado(cnx, filtroIdCliente, "N");
-        tablaActualizarCliente.setModel(modelo);
-    }//GEN-LAST:event_jTFIBuscadorAct1KeyReleased
-
-    private void jBIActualizarAct1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBIActualizarAct1ActionPerformed
-        List<String> camposInvalidos = new ArrayList<>();
-        List<String> atributos = new ArrayList<>();
-        ActualizarInventario actualizar = new ActualizarInventario();
-        String atributoActualizar = "idCliente";
-        String condicion = "'" + this.ciActualizar.getText() + "'";
-        String tabla = "cliente";
-        if (this.CIselect.isSelected()) {
-            if (!validadorCheck.validarCampoNoVacio(ciActualizar, "CI del cliente")) {
-                camposInvalidos.add("CI del cliente");
-            }
-        }
-
-        if (this.nombreSelect.isSelected()) {
-            if (!validadorCheck.validarCampoNoVacio(nombreActualizar, "Nombre del cliente")) {
-                camposInvalidos.add("Nombre del cliente");
-            }
-            atributos.add("nombres='" + nombreActualizar.getText() + "'");
-        }
-
-        if (this.apellidoSelect.isSelected()) {
-            if (!validadorCheck.validarCampoNoVacio(apellidoActualizar, "Apellido del cliente")) {
-                camposInvalidos.add("Apellido del cliente");
-            }
-            atributos.add("apellidos='" + apellidoActualizar.getText() + "'");
-        }
-
-        if (this.telefonoSelect.isSelected()) {
-            if (!validadorCheck.validarCampoNoVacio(telefonoActualizar, "Teléfono del cliente")) {
-                camposInvalidos.add("Teléfono del cliente");
-            }
-            atributos.add("telefonoCliente='" + telefonoActualizar.getText() + "'");
-        }
-
-        if (this.direccionSelect.isSelected()) {
-            if (!validadorCheck.validarCampoNoVacio(direccionActualizar, "Dirección del cliente")) {
-                camposInvalidos.add("Dirección del cliente");
-            }
-            atributos.add("direccionCliente='" + direccionActualizar.getText() + "'");
-        }
-
-        if (this.correoSelect.isSelected()) {
-            if (!validadorCheck.validarCampoNoVacio(correoActualizar, "Correo del cliente")) {
-                camposInvalidos.add("Correo del cliente");
-            }
-            atributos.add("correoElectronicoCliente='" + correoActualizar.getText() + "'");
-        }
-
-        if (!correoSelect.isSelected() && !direccionSelect.isSelected() && !telefonoSelect.isSelected()) {
-            JOptionPane.showMessageDialog(this, "Seleccione al menos un atributo a actualizar", "Error", JOptionPane.ERROR_MESSAGE);
-            return; // Detiene el registro si no se selecciona ningún atributo
-        }
-
-        if (!camposInvalidos.isEmpty()) {
-            String camposInvalidosStr = String.join(", ", camposInvalidos);
-            JOptionPane.showMessageDialog(this, "Los siguientes campos están vacíos: " + camposInvalidosStr, "Error", JOptionPane.ERROR_MESSAGE);
-
-            // Pinta de rojo los campos vacíos de forma parametrizada
-            for (String campoInvalido : camposInvalidos) {
-                switch (campoInvalido) {
-                    case "CI del cliente":
-                        validadorCheck.setColorFondoCampo(ciActualizar, new Color(255, 204, 204), errorActualizarCliente4);
-                        break;
-                    case "Nombre del cliente":
-                        validadorCheck.setColorFondoCampo(nombreActualizar, new Color(255, 204, 204), errorActualizarCliente5);
-                        break;
-                    case "Apellido del cliente":
-                        validadorCheck.setColorFondoCampo(apellidoActualizar, new Color(255, 204, 204), errorActualizarCliente6);
-                        break;
-                    case "Teléfono del cliente":
-                        validadorCheck.setColorFondoCampo(telefonoActualizar, new Color(255, 204, 204), errorActualizarCliente1);
-                        break;
-                    case "Dirección del cliente":
-                        validadorCheck.setColorFondoCampo(direccionActualizar, new Color(255, 204, 204), errorActualizarCliente2);
-                        break;
-                    case "Correo del cliente":
-                        validadorCheck.setColorFondoCampo(correoActualizar, new Color(255, 204, 204), errorActualizarCliente3);
-                        break;
-
-                    // Puedes agregar más casos aquí para otros campos si es necesario
-                }
-            }
-            return; // Detiene el registro si hay campos inválidos
-        }
-        String atributosActualizacion = String.join(", ", atributos);
-        if (!telfCliente1 && !dirCliente1 && !correoElectronico1) {
-            JOptionPane.showMessageDialog(this, "Ha ocurrido un error al intentar actualizar. "
-                    + "\nPor favor, verifica los campos e inténtalo nuevamente.", "Error en el registro", JOptionPane.ERROR_MESSAGE);
-        } else {
-            actualizar.actualizarDatosF(this.cnx, atributoActualizar, condicion, tabla, atributosActualizacion);
-            String filtroIdCliente = jTFIBuscadorAct1.getText();
-            CreadorTablas creadorTablas = new CreadorTablas();
-            DefaultTableModel modelo = creadorTablas.mostrarClienteParametrizado(cnx, filtroIdCliente, "N");
-            tablaActualizarCliente.setModel(modelo);
-            JTextField[] camposParaLimpiar = {ciActualizar, nombreActualizar, apellidoActualizar, telefonoActualizar, direccionActualizar, correoActualizar};
-            validadorCheck.limpiarCampos(camposParaLimpiar);
-        }
-    }//GEN-LAST:event_jBIActualizarAct1ActionPerformed
-
-    private void tablaActualizarClienteMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tablaActualizarClienteMouseClicked
-        int filaSeleccionada = tablaActualizarCliente.getSelectedRow();
-        boolean foundColumn = false;
-        if (filaSeleccionada != -1) {
-            DefaultTableModel model = (DefaultTableModel) tablaActualizarCliente.getModel();
-            for (int i = 0; i < model.getColumnCount(); i++) {
-                String columnName = model.getColumnName(i);
-                if (columnName.equals("ID Cliente")
-                        || columnName.equals("Nombres") || columnName.equals("Apellidos")
-                        || columnName.equals("Teléfono movíl")
-                        || columnName.equals("Dirección") || columnName.equals("Nacionalidad")
-                        || columnName.equals("Tipo") || columnName.equals("Correo electronico")) {
-                    String idCliente = getValueAtSelectedRow(model, filaSeleccionada, "ID Cliente");
-                    String nombreClienteA = getValueAtSelectedRow(model, filaSeleccionada, "Nombres");
-                    String apellido = getValueAtSelectedRow(model, filaSeleccionada, "Apellidos");
-                    String telefonoClienteA = getValueAtSelectedRow(model, filaSeleccionada, "Teléfono movíl");
-                    String direccionClienteA = getValueAtSelectedRow(model, filaSeleccionada, "Dirección");
-                    String esExtranjero = getValueAtSelectedRow(model, filaSeleccionada, "Nacionalidad");
-                    String tipo = getValueAtSelectedRow(model, filaSeleccionada, "Tipo");
-                    String correoClienteA = getValueAtSelectedRow(model, filaSeleccionada, "Correo electronico");
-
-                    if (esExtranjero.equals("Extranjero")) {  // Supongo que "true" indica extranjero
-                        jCBNacionalExtranjero2.setSelectedItem("Extranjero");
-                    } else {
-                        if (esExtranjero.equals("Nacional")) {  // Supongo que "true" indica extranjero
-                            jCBNacionalExtranjero2.setSelectedItem("Nacional");
-                        }
-                    }
-                    if (tipo.equals("Jurídico")) {  // Supongo que "true" indica extranjero
-                        jCJuridicoNatural1.setSelectedItem("Jurídico");
-                    } else {
-                        if (tipo.equals("Natural")) {  // Supongo que "true" indica extranjero
-                            jCJuridicoNatural1.setSelectedItem("Natural");
-                        }
-                    }
-                    validadorCheck.actualizarCampoSeleccionado(CIselect, idCliente, ciActualizar, errorActualizarCliente4);
-                    validadorCheck.actualizarCampoSeleccionado(nombreSelect, nombreClienteA, nombreActualizar, errorActualizarCliente5);
-                    validadorCheck.actualizarCampoSeleccionado(apellidoSelect, apellido, apellidoActualizar, errorActualizarCliente6);
-                    validadorCheck.actualizarCampoSeleccionado(telefonoSelect, telefonoClienteA, telefonoActualizar, errorActualizarCliente1);
-                    validadorCheck.actualizarCampoSeleccionado(direccionSelect, direccionClienteA, direccionActualizar, errorActualizarCliente2);
-                    validadorCheck.actualizarCampoSeleccionado(correoSelect, correoClienteA, correoActualizar, errorActualizarCliente3);
-                    Boolean[] valoresItem = {telfCliente1, dirCliente1, correoElectronico1};
-                    valoresItem = validadorCheck.cambiarValoresVerdadFinal(valoresItem);
-                    telfCliente1 = valoresItem[0];
-                    dirCliente1 = valoresItem[1];
-                    correoElectronico1 = valoresItem[2];
-                    foundColumn = true;
-                    break; // Sale del bucle una vez que se haya encontrado una columna válida                 
-                }
-            }
-        }
-    }//GEN-LAST:event_tablaActualizarClienteMouseClicked
 
     private void menuAdministracionMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_menuAdministracionMouseClicked
         contenido.show(panelContent, "card7");
@@ -5961,7 +4865,7 @@ public class JFMenu extends javax.swing.JFrame {
         correoCli.setText(correo);
     }
 
-    private void fillClientFields1(String id, String nombre, String apellido,
+    /*private void fillClientFields1(String id, String nombre, String apellido,
             String telefono, String direccion,
             String esExtranjero, String tipo, String correo) {
         if (esExtranjero.equals("Extranjero")) {  // Supongo que "true" indica extranjero
@@ -5985,8 +4889,7 @@ public class JFMenu extends javax.swing.JFrame {
         jTFTelefonoR.setText(telefono);
         jTFDireccionR.setText(direccion);
         correoCli2.setText(correo);
-    }
-
+    }*/
 
     private void jBGenerarFacturaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBGenerarFacturaActionPerformed
         int numeroFactura = Integer.parseInt(jTFnumerofactura.getText());
@@ -6038,38 +4941,6 @@ public class JFMenu extends javax.swing.JFrame {
     private void correoCliFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_correoCliFocusLost
         correoElectronico = validarRegistroF.camposDeRegistros(correoCli, jError12, "c");
     }//GEN-LAST:event_correoCliFocusLost
-
-    private void jCJuridicoNaturalActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCJuridicoNaturalActionPerformed
-        jTFCIRegistrarC.setText("");
-        String seleccion = (String) jCJuridicoNatural.getSelectedItem();
-        validarCamposJuridicoC();
-        if ("Jurídico".equals(seleccion)) {
-            jLTipoCli.setText("RUC");
-        } else if ("Natural".equals(seleccion)) {
-            String seleccionExtranjero = (String) jCBNacionalExtranjero.getSelectedItem();
-            if ("Nacional".equals(seleccionExtranjero)) {
-                jLTipoCli.setText("CI");
-            } else if ("Extranjero".equals(seleccionExtranjero)) {
-                jLTipoCli.setText("Pasaporte");
-            }
-        }
-    }//GEN-LAST:event_jCJuridicoNaturalActionPerformed
-
-    private void jCBNacionalExtranjeroActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCBNacionalExtranjeroActionPerformed
-        jTFCIRegistrarC.setText("");
-        String seleccion2 = (String) jCJuridicoNatural.getSelectedItem();
-        String seleccion = (String) jCBNacionalExtranjero.getSelectedItem();
-        validarCamposJuridicoC();
-        if ("Natural".equals(seleccion2)) {
-            if ("Nacional".equals(seleccion)) {
-                jLTipoCli.setText("CI");
-            } else if ("Extranjero".equals(seleccion)) {
-                jLTipoCli.setText("Pasaporte");
-            }
-        } else if ("Jurídico".equals(seleccion2)) {
-            jLTipoCli.setText("RUC");
-        }
-    }//GEN-LAST:event_jCBNacionalExtranjeroActionPerformed
 
     private void jCBNacionalExtranjero1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCBNacionalExtranjero1ActionPerformed
         jTCIDelCliente.setText("");
@@ -6131,17 +5002,6 @@ public class JFMenu extends javax.swing.JFrame {
             jTCIDelCliente.setEnabled(false); // Deshabilitar el campo si las selecciones no son válidas
         } else {
             jTCIDelCliente.setEnabled(true); // Habilitar el campo si las selecciones son válidas
-        }
-    }
-
-    private void validarCamposJuridicoC() {
-        String nacionalExtranjero = jCBNacionalExtranjero.getSelectedItem().toString();
-        String naturalJuridico = jCJuridicoNatural.getSelectedItem().toString();
-
-        if (nacionalExtranjero.equals("Selecciona") || naturalJuridico.equals("Natural o Jurídico")) {
-            jTFCIRegistrarC.setEnabled(false); // Deshabilitar el campo si las selecciones no son válidas
-        } else {
-            jTFCIRegistrarC.setEnabled(true); // Habilitar el campo si las selecciones son válidas
         }
     }
 
@@ -6351,63 +5211,6 @@ public class JFMenu extends javax.swing.JFrame {
         String setIva = AdminIVA.getText();
         IVAText.setText(setIva);
     }//GEN-LAST:event_AdminIVAKeyReleased
-
-    private void jBRegistarClienteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBRegistarClienteActionPerformed
-        JTextField[] campos = {jTFCIRegistrarC, jTFNombresR, jTFApellidosR, correoCli2,
-            jTFTelefonoR, jTFDireccionR};
-        JLabel[] labels = {errorc1, errorc2, errorc3, errorc4, errorc5, errorc6};
-        Boolean[] booleanosCliente = {documentoCliente, nombreCliente, apellidoCliente, correoCliente,
-            telefonoCliente, direccionCliente};
-        String[] nombresCampos = {"documento", "nombre", "apellido", "correo", "teléfono", "dirección"};
-        List<String> errores = validadorCheck.validarCamposLista(campos, booleanosCliente, labels, nombresCampos);
-        errores.addAll(validadorCheck.validarCamposVaciosLista(campos, booleanosCliente, labels, nombresCampos));
-        if (!errores.isEmpty()) {
-            StringBuilder mensajeError = new StringBuilder("Se encontraron los siguientes errores:\n");
-            for (String error : errores) {
-                mensajeError.append("- ").append(error).append("\n");
-            }
-            JOptionPane.showMessageDialog(null, mensajeError.toString(), "Errores", JOptionPane.ERROR_MESSAGE);
-        } else {
-            Object[] options = {"Si", "No"};
-            int opcion = JOptionPane.showOptionDialog(null, "¿Deseas registrar los datos?", "Confirmación",
-                    JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE, null, options, options[0]);
-
-            if (opcion == JOptionPane.YES_OPTION) {
-                String esExtranjero = (String) jCBNacionalExtranjero.getSelectedItem();
-                String nombresCliente = jTFNombresR.getText();
-                String apellidosCliente = jTFApellidosR.getText();
-                String telefonoClientes = jTFTelefonoR.getText();
-                String direccionClientes = jTFDireccionR.getText();
-                String tipo = (String) jCJuridicoNatural.getSelectedItem();
-                String correoDelCli = correoCli2.getText();
-                String idCliente = jTFCIRegistrarC.getText();
-
-                try {
-                    RegistrarDatosFactura.insertarDatosCliente(cnx, idCliente, nombresCliente,
-                            apellidosCliente, telefonoClientes,
-                            direccionClientes, esExtranjero, tipo, correoDelCli);
-                    CreadorTablas creartabla = new CreadorTablas();
-                    DefaultTableModel modelo = creartabla.mostrarClienteParametrizado(cnx, idCliente, "N");
-                    jTablaClientesR.setModel(modelo);
-                } catch (SQLException ex) {
-                    Logger.getLogger(JFMenu.class.getName()).log(Level.SEVERE, null, ex);
-                }
-                JOptionPane.showMessageDialog(null, "Datos registrados correctamente");
-            } else {
-                // El usuario seleccionó "No", puedes realizar alguna otra acción o mostrar un mensaje adecuado
-            }
-        }
-    }//GEN-LAST:event_jBRegistarClienteActionPerformed
-
-    private void jTFCIRegistrarCKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFCIRegistrarCKeyReleased
-        String filtroIdCliente = jTFCIRegistrarC.getText();
-        CreadorTablas creadorTablas = new CreadorTablas();
-        DefaultTableModel modelo = creadorTablas.mostrarClienteParametrizado(cnx, filtroIdCliente, "N"); // Llamas al método con el filtro
-        jTablaClientesR.setModel(modelo); // Actualizas la tabla con el nuevo modelo 
-        String tipoCliente = (String) jCBNacionalExtranjero.getSelectedItem();
-        String natularJuridico = (String) jCJuridicoNatural.getSelectedItem();
-        documentoCliente = validarRegistroF.camposCliente(jTCIDelCliente, jError7, tipoCliente, natularJuridico, "a");
-    }//GEN-LAST:event_jTFCIRegistrarCKeyReleased
 
     private void BRegistrarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_BRegistrarActionPerformed
         boolean registro = true;
@@ -6626,52 +5429,6 @@ public class JFMenu extends javax.swing.JFrame {
         telefonoNeg = true;
     }
 
-    public void cambiarValoresNegoP() {
-        nombreNegocioValido2 = true;
-        estadoPagoP2 = true;
-        rucV2 = true;
-        dirNeg2 = true;
-        telefonoNeg2 = true;
-    }
-
-    public void cambiarValoresClienteP() {
-        rucCliente2 = true;
-        nombreCFactura2 = true;
-        apellidoCFactura2 = true;
-        telfCliente2 = true;
-        dirCliente2 = true;
-        correoElectronico2 = true;
-    }
-
-    public void cambiarValoresCliente() {
-        rucCliente = true;
-        nombreCFactura = true;
-        apellidoCFactura = true;
-        telfCliente = true;
-        dirCliente = true;
-        correoElectronico = true;
-    }
-
-    public void cambiarValoresVerdad1() {
-        telfCliente1 = true;
-        dirCliente1 = true;
-        correoElectronico1 = true;
-    }
-
-
-    private void jTFNombresRKeyTyped(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFNombresRKeyTyped
-        int maxLength = 40; // Límite de longitud de caracteres
-        if (jTFNombresR.getText().length() >= maxLength) {
-            evt.consume();
-        }
-    }//GEN-LAST:event_jTFNombresRKeyTyped
-
-    private void jTFApellidosRKeyTyped(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFApellidosRKeyTyped
-        int maxLength = 64; // Límite de longitud de caracteres
-        if (jTFApellidosR.getText().length() >= maxLength) {
-            evt.consume();
-        }
-    }//GEN-LAST:event_jTFApellidosRKeyTyped
 
     private void jTNombreClienteFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTNombreClienteFocusLost
         nombreCFactura = validarRegistroF.camposDeRegistros(jTNombreCliente, jError8, "n");
@@ -6703,44 +5460,6 @@ public class JFMenu extends javax.swing.JFrame {
     private void ClaveEmpleadoUsuarioKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_ClaveEmpleadoUsuarioKeyReleased
         validarClave(ClaveEmpleadoUsuario, "2");
     }//GEN-LAST:event_ClaveEmpleadoUsuarioKeyReleased
-
-    private void telefonoActualizarKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_telefonoActualizarKeyReleased
-        String tipoCliente = (String) jCBNacionalExtranjero2.getSelectedItem();
-        telfCliente1 = validarRegistroF.camposCliente(telefonoActualizar, errorActualizarCliente1, tipoCliente, "Ninguno", "b");
-    }//GEN-LAST:event_telefonoActualizarKeyReleased
-
-    private void direccionActualizarKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_direccionActualizarKeyReleased
-        dirCliente1 = validarRegistroF.camposDeRegistros(direccionActualizar, errorActualizarCliente2, "d");
-    }//GEN-LAST:event_direccionActualizarKeyReleased
-
-    private void correoActualizarKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_correoActualizarKeyReleased
-        correoElectronico1 = validarRegistroF.camposDeRegistros(correoActualizar, errorActualizarCliente3, "c");
-    }//GEN-LAST:event_correoActualizarKeyReleased
-
-    private void telefonoActualizarFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_telefonoActualizarFocusLost
-        String tipoCliente = (String) jCBNacionalExtranjero2.getSelectedItem();
-        telfCliente1 = validarRegistroF.camposCliente(telefonoActualizar, errorActualizarCliente1, tipoCliente, "Ninguno", "b");
-    }//GEN-LAST:event_telefonoActualizarFocusLost
-
-    private void direccionActualizarFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_direccionActualizarFocusLost
-        dirCliente1 = validarRegistroF.camposDeRegistros(direccionActualizar, errorActualizarCliente2, "d");
-    }//GEN-LAST:event_direccionActualizarFocusLost
-
-    private void correoActualizarFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_correoActualizarFocusLost
-        correoElectronico1 = validarRegistroF.camposDeRegistros(correoActualizar, errorActualizarCliente3, "c");
-    }//GEN-LAST:event_correoActualizarFocusLost
-
-    private void correoSelectActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_correoSelectActionPerformed
-        validadorCheck.actualizarCampo(correoSelect, correoActualizar, correoElectronico1, errorActualizarCliente3);
-    }//GEN-LAST:event_correoSelectActionPerformed
-
-    private void direccionSelectActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_direccionSelectActionPerformed
-        validadorCheck.actualizarCampo(direccionSelect, direccionActualizar, dirCliente1, errorActualizarCliente2);
-    }//GEN-LAST:event_direccionSelectActionPerformed
-
-    private void telefonoSelectActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_telefonoSelectActionPerformed
-        validadorCheck.actualizarCampo(telefonoSelect, telefonoActualizar, telfCliente1, errorActualizarCliente1);
-    }//GEN-LAST:event_telefonoSelectActionPerformed
 
     private void jTFIDemployeFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTFIDemployeFocusLost
         String idCedulaEmpleado = jTFIDemploye.getText();
@@ -7273,17 +5992,6 @@ public class JFMenu extends javax.swing.JFrame {
         validadorCheck.actualizarCampo(jCheckBox8, jTATelefonoDespachador, telefonoProve1, errorProveedores7);
     }//GEN-LAST:event_jCheckBox8ActionPerformed
 
-    private void jTFCIRegistrarCFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTFCIRegistrarCFocusLost
-        String idCliente = jTFCIRegistrarC.getText();
-        RegistrarDatosFactura registrar = new RegistrarDatosFactura();
-        if (registrar.clienteYaRegistrado(cnx, idCliente)) {
-            JOptionPane.showMessageDialog(this, "Cliente ya registrado", "Error", JOptionPane.ERROR_MESSAGE);
-        }
-        String tipoCliente = (String) jCBNacionalExtranjero.getSelectedItem();
-        String natularJuridico = (String) jCJuridicoNatural.getSelectedItem();
-        documentoCliente = validarRegistroF.camposCliente(jTCIDelCliente, jError7, tipoCliente, natularJuridico, "a");
-    }//GEN-LAST:event_jTFCIRegistrarCFocusLost
-
     private void claveUsuariocheckActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_claveUsuariocheckActionPerformed
         if (claveUsuariocheck.isSelected()) {
             ClaveEmpleadoUsuario.setEnabled(true);
@@ -7336,58 +6044,13 @@ public class JFMenu extends javax.swing.JFrame {
         apellidoCFactura = validarRegistroF.camposDeRegistros(jTFApeliidosCliente, jError9, "n");
     }//GEN-LAST:event_jTFApeliidosClienteKeyReleased
 
-    private void jTFNombresRKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFNombresRKeyReleased
-        nombreCliente = validarRegistroF.camposDeRegistros(jTFNombresR, errorc2, "n");
-    }//GEN-LAST:event_jTFNombresRKeyReleased
-
-    private void jTFNombresRFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTFNombresRFocusLost
-        nombreCliente = validarRegistroF.camposDeRegistros(jTFNombresR, errorc2, "n");
-    }//GEN-LAST:event_jTFNombresRFocusLost
-
-    private void jTFApellidosRKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFApellidosRKeyReleased
-        apellidoCliente = validarRegistroF.camposDeRegistros(jTFApellidosR, errorc3, "n");
-    }//GEN-LAST:event_jTFApellidosRKeyReleased
-
-    private void jTFApellidosRFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTFApellidosRFocusLost
-        apellidoCliente = validarRegistroF.camposDeRegistros(jTFApellidosR, errorc3, "n");
-    }//GEN-LAST:event_jTFApellidosRFocusLost
-
-    private void correoCli2KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_correoCli2KeyReleased
-        correoCliente = validarRegistroF.camposDeRegistros(correoCli2, errorc4, "c");
-    }//GEN-LAST:event_correoCli2KeyReleased
-
-    private void correoCli2FocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_correoCli2FocusLost
-        correoCliente = validarRegistroF.camposDeRegistros(correoCli2, errorc4, "c");
-    }//GEN-LAST:event_correoCli2FocusLost
-
-    private void jTFTelefonoRFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTFTelefonoRFocusLost
-        String tipoCliente = (String) jCBNacionalExtranjero.getSelectedItem();
-        telefonoCliente = validarRegistroF.camposCliente(jTFTelefonoR, errorc5, tipoCliente, "Ninguno", "b");
-    }//GEN-LAST:event_jTFTelefonoRFocusLost
-
-    private void jTFTelefonoRKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFTelefonoRKeyReleased
-        String tipoCliente = (String) jCBNacionalExtranjero.getSelectedItem();
-        telefonoCliente = validarRegistroF.camposCliente(jTFTelefonoR, errorc5, tipoCliente, "Ninguno", "b");
-    }//GEN-LAST:event_jTFTelefonoRKeyReleased
-
-    private void jTFDireccionRKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTFDireccionRKeyReleased
-        direccionCliente = validarRegistroF.camposDeRegistros(jTFDireccionR, errorc6, "d");
-    }//GEN-LAST:event_jTFDireccionRKeyReleased
-
-    private void jTFDireccionRFocusLost(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_jTFDireccionRFocusLost
-        direccionCliente = validarRegistroF.camposDeRegistros(jTFDireccionR, errorc6, "d");
-    }//GEN-LAST:event_jTFDireccionRFocusLost
-
-
-    
 
     private void menuIncidentesMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_menuIncidentesMouseClicked
         contenido.show(panelContent, "card6");
-        //Empleados empleados = new Empleados(cnx);
-        //DefaultTableModel modelo = empleados.obtenerModeloTablaEmpleadosActualizar();
-        //jTableEmpleadosAcutalizar.setModel(modelo);
         cambiarSeccionMenu(5);
         menuIncidentes.setBackground(Color.decode("#494848"));
+        JFrame incidente = new JFIncidente(cnx);
+        VentanaManager.getInstance().mostrarVentana("inventario", incidente);
     }//GEN-LAST:event_menuIncidentesMouseClicked
 
     private void menuIncidentesMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_menuIncidentesMouseEntered
@@ -7399,93 +6062,6 @@ public class JFMenu extends javax.swing.JFrame {
         menuIncidentes.setBackground(Color.decode("#292728"));
         menuIncidentes.setOpaque(true);
     }//GEN-LAST:event_menuIncidentesMouseExited
-
-    private void jTPEmpleados1MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jTPEmpleados1MouseClicked
-
-    }//GEN-LAST:event_jTPEmpleados1MouseClicked
-
-    private void SeleccionIncidentesCBActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_SeleccionIncidentesCBActionPerformed
-
-    }//GEN-LAST:event_SeleccionIncidentesCBActionPerformed
-
-    private void IDIncidentesTFActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_IDIncidentesTFActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_IDIncidentesTFActionPerformed
-
-    private void IDPIncidentesTFKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_IDPIncidentesTFKeyReleased
-        String idPaqueteStr = IDPIncidentesTF.getText().trim();
-        DefaultTableModel modelo = new CreadorTablas().mostrarVistaDatosPaquete(cnx, idPaqueteStr);
-        jTIncidentes.setModel(modelo);
-    }//GEN-LAST:event_IDPIncidentesTFKeyReleased
-
-    private void RegistrarIncidentesButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_RegistrarIncidentesButtonActionPerformed
-        int idIncidente = Integer.parseInt(IDIncidentesTF.getText());
-        int idPaquete = Integer.parseInt(IDPIncidentesTF.getText());
-        String tipoIncidente = SeleccionIncidentesCB.getSelectedItem().toString();
-        String descripcion = RazonIncidentesTA.getText();
-
-        // Crear instancia de GestorIncidentes
-        Inventario inventario = new Inventario(); // Asegúrate de tener una instancia de Inventario
-        GestorIncidentes gestorIncidentes = new GestorIncidentes(inventario, cnx);
-
-        // Crear el incidente según el tipo
-        Incidente incidente = null;
-        switch (tipoIncidente) {
-            case "Daño en el Paquete":
-                incidente = new DanioPaquete(descripcion, idPaquete, idIncidente, cnx);
-                break;
-            case "Error de Dirección":
-                incidente = new ErrorDireccion(descripcion, idPaquete, idIncidente, cnx);
-                break;
-            case "Paquete Perdido":
-                incidente = new PaquetePerdido(descripcion, idPaquete, idIncidente, cnx);
-                break;
-            case "Rechazo Entrega":
-                incidente = new RechazoEntrega(descripcion, idPaquete, idIncidente, cnx);
-                break;
-            default:
-                throw new IllegalArgumentException("Tipo de incidente desconocido: " + tipoIncidente);
-        }
-
-        // Registrar el incidente
-        gestorIncidentes.crearIncidente(incidente, idPaquete);
-        contadorIncidentes();
-    }//GEN-LAST:event_RegistrarIncidentesButtonActionPerformed
-
-    private void jCJuridicoNatural1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCJuridicoNatural1ActionPerformed
-        //jTFCIRegistrarC.setText("");
-        String seleccion = (String) jCJuridicoNatural1.getSelectedItem();
-        if ("Jurídico".equals(seleccion)) {
-            CIselect.setText("RUC");
-            documentoActualizar.setText("RUC");
-        } else if ("Natural".equals(seleccion)) {
-            String seleccionExtranjero = (String) jCBNacionalExtranjero.getSelectedItem();
-            if ("Nacional".equals(seleccionExtranjero)) {
-                CIselect.setText("CI");
-                documentoActualizar.setText("CI");
-            } else if ("Extranjero".equals(seleccionExtranjero)) {
-                CIselect.setText("Pasaporte");
-                documentoActualizar.setText("Pasaporte");
-            }
-        }
-    }//GEN-LAST:event_jCJuridicoNatural1ActionPerformed
-
-    private void jCBNacionalExtranjero2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCBNacionalExtranjero2ActionPerformed
-        String seleccion2 = (String) jCJuridicoNatural1.getSelectedItem();
-        String seleccion = (String) jCBNacionalExtranjero2.getSelectedItem();
-        if ("Natural".equals(seleccion2)) {
-            if ("Nacional".equals(seleccion)) {
-                CIselect.setText("CI");
-                documentoActualizar.setText("CI");
-            } else if ("Extranjero".equals(seleccion)) {
-                CIselect.setText("Pasaporte");
-                documentoActualizar.setText("Pasaporte");
-            }
-        } else if ("Jurídico".equals(seleccion2)) {
-            CIselect.setText("RUC");
-            documentoActualizar.setText("RUC");
-        }
-    }//GEN-LAST:event_jCBNacionalExtranjero2ActionPerformed
 
     private void limpiarTabla(JTable tabla) {
         DefaultTableModel modelo = (DefaultTableModel) tabla.getModel();
@@ -7760,7 +6336,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JButton BRegistrar;
     private javax.swing.JComboBox<String> CBSexoActualizado;
     private javax.swing.JTextField CIEmpleadoActualizar;
-    private javax.swing.JCheckBox CIselect;
     private javax.swing.JButton CambiarEstadoF;
     private javax.swing.JTextField Cargo;
     private javax.swing.JTextField CedulaEmpleado;
@@ -7775,8 +6350,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JTextField CorreoElectronico;
     private com.toedter.calendar.JDateChooser FechaNacimientoEmpleado;
     private javax.swing.JPanel Home;
-    private javax.swing.JTextField IDIncidentesTF;
-    private javax.swing.JTextField IDPIncidentesTF;
     private javax.swing.JTextField IVAText;
     private javax.swing.JPanel JPAdministración;
     private javax.swing.JPanel JPConductores;
@@ -7787,9 +6360,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JPanel JPRemitente;
     private javax.swing.JTextField NombreEmpleado;
     private javax.swing.JPanel PanelHome;
-    private javax.swing.JTextArea RazonIncidentesTA;
-    private javax.swing.JButton RegistrarIncidentesButton;
-    private javax.swing.JComboBox<String> SeleccionIncidentesCB;
     private javax.swing.JComboBox<String> SexoEmpleado;
     private javax.swing.JTextField TFApellidosActualizado;
     private javax.swing.JTextField TFCargoActualizado;
@@ -7802,28 +6372,13 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JTextField TFTPersonalActualizado;
     private javax.swing.JTextField TelefonoConvencional;
     private javax.swing.JTextField TelefonoPersonal;
-    private javax.swing.JTextField apellidoActualizar;
-    private javax.swing.JCheckBox apellidoSelect;
     private javax.swing.JButton btnExit;
     private javax.swing.JButton btnHelp;
     private javax.swing.JTextField campoRespuesta;
-    private javax.swing.JTextField ciActualizar;
     private javax.swing.JCheckBox claveUsuariocheck;
-    private javax.swing.JTextField correoActualizar;
     private javax.swing.JTextField correoCli;
-    private javax.swing.JTextField correoCli2;
-    private javax.swing.JCheckBox correoSelect;
-    private javax.swing.JTextField direccionActualizar;
     private javax.swing.JTextField direccionEmpelado;
-    private javax.swing.JCheckBox direccionSelect;
-    private javax.swing.JLabel documentoActualizar;
     private javax.swing.JLabel documentoActualizar3;
-    private javax.swing.JLabel errorActualizarCliente1;
-    private javax.swing.JLabel errorActualizarCliente2;
-    private javax.swing.JLabel errorActualizarCliente3;
-    private javax.swing.JLabel errorActualizarCliente4;
-    private javax.swing.JLabel errorActualizarCliente5;
-    private javax.swing.JLabel errorActualizarCliente6;
     private javax.swing.JLabel errorEmpleados1;
     private javax.swing.JLabel errorProveedores1;
     private javax.swing.JLabel errorProveedores2;
@@ -7833,12 +6388,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JLabel errorProveedores6;
     private javax.swing.JLabel errorProveedores7;
     private javax.swing.JLabel errorProveedores8;
-    private javax.swing.JLabel errorc1;
-    private javax.swing.JLabel errorc2;
-    private javax.swing.JLabel errorc3;
-    private javax.swing.JLabel errorc4;
-    private javax.swing.JLabel errorc5;
-    private javax.swing.JLabel errorc6;
     private javax.swing.JTextField idItemFactura;
     private javax.swing.JButton jBAgregarProducto;
     private javax.swing.JButton jBAgregarProductoProveedores;
@@ -7846,12 +6395,10 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JButton jBConsultarInventario3;
     private javax.swing.JButton jBGenerarFactura;
     private javax.swing.JButton jBIActualizarAct;
-    private javax.swing.JButton jBIActualizarAct1;
     private javax.swing.JButton jBIActualizarAct2;
     private javax.swing.JButton jBQuitarProducto;
     private javax.swing.JButton jBQuitarProductoProveedores;
     private javax.swing.JButton jBQuitarProductoProveedores1;
-    private javax.swing.JButton jBRegistarCliente;
     private javax.swing.JButton jBRegistarUser;
     private javax.swing.JButton jBRegistrarFactura;
     private javax.swing.JButton jBRegistrarProovedor;
@@ -7862,9 +6409,7 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JCheckBox jCApellidoActualizar;
     private javax.swing.JComboBox<String> jCBEstadoPago;
     private javax.swing.JComboBox<String> jCBJuridicoNatural1;
-    private javax.swing.JComboBox<String> jCBNacionalExtranjero;
     private javax.swing.JComboBox<String> jCBNacionalExtranjero1;
-    private javax.swing.JComboBox<String> jCBNacionalExtranjero2;
     private javax.swing.JComboBox<String> jCBPreguntaControl;
     private javax.swing.JComboBox<String> jCBseleccionRol;
     private javax.swing.JComboBox<String> jCBuscarProveedorConsultar2;
@@ -7873,8 +6418,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JCheckBox jCCorreoActualizar;
     private javax.swing.JCheckBox jCDireccionActualizar;
     private javax.swing.JCheckBox jCFechaActualizar;
-    private javax.swing.JComboBox<String> jCJuridicoNatural;
-    private javax.swing.JComboBox<String> jCJuridicoNatural1;
     private javax.swing.JCheckBox jCNombreActualizar;
     private javax.swing.JCheckBox jCSexoActualizar;
     private javax.swing.JCheckBox jCTelefonoCActualizar;
@@ -7917,7 +6460,7 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JLabel jLErrorEmpleado7;
     private javax.swing.JLabel jLErrorEmpleado8;
     private javax.swing.JLabel jLErrorEmpleado9;
-    private javax.swing.JLabel jLTipoCli;
+    private javax.swing.JLabel jLInicio;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel107;
     private javax.swing.JLabel jLabel109;
@@ -7925,11 +6468,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JLabel jLabel113;
     private javax.swing.JLabel jLabel114;
     private javax.swing.JLabel jLabel117;
-    private javax.swing.JLabel jLabel132;
-    private javax.swing.JLabel jLabel133;
-    private javax.swing.JLabel jLabel134;
-    private javax.swing.JLabel jLabel135;
-    private javax.swing.JLabel jLabel136;
     private javax.swing.JLabel jLabel15;
     private javax.swing.JLabel jLabel16;
     private javax.swing.JLabel jLabel17;
@@ -7973,16 +6511,12 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JLabel jLabel57;
     private javax.swing.JLabel jLabel58;
     private javax.swing.JLabel jLabel59;
-    private javax.swing.JLabel jLabel6;
     private javax.swing.JLabel jLabel64;
     private javax.swing.JLabel jLabel65;
-    private javax.swing.JLabel jLabel66;
     private javax.swing.JLabel jLabel67;
     private javax.swing.JLabel jLabel68;
     private javax.swing.JLabel jLabel69;
     private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel71;
-    private javax.swing.JLabel jLabel73;
     private javax.swing.JLabel jLabel74;
     private javax.swing.JLabel jLabel75;
     private javax.swing.JLabel jLabel76;
@@ -7992,44 +6526,30 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JLabel jLabel8;
     private javax.swing.JLabel jLabel84;
     private javax.swing.JLabel jLabel85;
-    private javax.swing.JLabel jLabel87;
-    private javax.swing.JLabel jLabel89;
     private javax.swing.JLabel jLabel9;
     private javax.swing.JPanel jPAE;
-    private javax.swing.JPanel jPAE1;
-    private javax.swing.JPanel jPActualizarClientes;
     private javax.swing.JPanel jPActualizarProovedores;
     private javax.swing.JPanel jPCE;
-    private javax.swing.JPanel jPCambiarEstado;
     private javax.swing.JPanel jPCambiarEstadoP;
-    private javax.swing.JTabbedPane jPClientes;
     private javax.swing.JPanel jPDatosCliente;
     private javax.swing.JPanel jPDatosEmpleados;
     private javax.swing.JPanel jPDatosFactura;
     private javax.swing.JPanel jPDatosNegocio;
     private javax.swing.JPanel jPDatosRecuperadosEmpleados;
     private javax.swing.JPanel jPEE;
-    private javax.swing.JPanel jPEE1;
     private javax.swing.JPanel jPEmpleadosTab;
     private javax.swing.JTabbedPane jPGP;
-    private javax.swing.JPanel jPIncidentes;
     private javax.swing.JPanel jPPA;
-    private javax.swing.JPanel jPPA1;
     private javax.swing.JPanel jPPA2;
     private javax.swing.JPanel jPPC;
-    private javax.swing.JPanel jPPC1;
     private javax.swing.JPanel jPPC2;
-    private javax.swing.JPanel jPPCE1;
     private javax.swing.JPanel jPPCE2;
     private javax.swing.JPanel jPPR;
-    private javax.swing.JPanel jPPR1;
     private javax.swing.JPanel jPPR2;
     private javax.swing.JPanel jPProducto;
     private javax.swing.JPanel jPProvedoresActualizar;
     private javax.swing.JPanel jPRE;
-    private javax.swing.JPanel jPRE1;
     private javax.swing.JPanel jPanel1;
-    private javax.swing.JPanel jPanel10;
     private javax.swing.JPanel jPanel11;
     private javax.swing.JPanel jPanel12;
     private javax.swing.JPanel jPanel13;
@@ -8047,13 +6567,8 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JPanel jPanel3;
     private javax.swing.JPanel jPanel31;
     private javax.swing.JPanel jPanel33;
-    private javax.swing.JPanel jPanel37;
-    private javax.swing.JPanel jPanel38;
-    private javax.swing.JPanel jPanel40;
     private javax.swing.JPanel jPanel42;
-    private javax.swing.JPanel jPanel43;
     private javax.swing.JPanel jPanel45;
-    private javax.swing.JPanel jPanel46;
     private javax.swing.JPanel jPanel47;
     private javax.swing.JPanel jPanel49;
     private javax.swing.JPanel jPanel50;
@@ -8064,8 +6579,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JScrollPane jScrollPane11;
     private javax.swing.JScrollPane jScrollPane12;
     private javax.swing.JScrollPane jScrollPane13;
-    private javax.swing.JScrollPane jScrollPane16;
-    private javax.swing.JScrollPane jScrollPane18;
     private javax.swing.JScrollPane jScrollPane2;
     private javax.swing.JScrollPane jScrollPane20;
     private javax.swing.JScrollPane jScrollPane21;
@@ -8074,9 +6587,7 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JScrollPane jScrollPane24;
     private javax.swing.JScrollPane jScrollPane25;
     private javax.swing.JScrollPane jScrollPane26;
-    private javax.swing.JScrollPane jScrollPane28;
     private javax.swing.JScrollPane jScrollPane3;
-    private javax.swing.JScrollPane jScrollPane4;
     private javax.swing.JScrollPane jScrollPane5;
     private javax.swing.JScrollPane jScrollPane6;
     private javax.swing.JSeparator jSeparator1;
@@ -8095,27 +6606,20 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JComboBox<String> jTFAdminRol;
     private javax.swing.JTextField jTFApeliidosCliente;
     private javax.swing.JTextField jTFApellidoEmpleado;
-    private javax.swing.JTextField jTFApellidosR;
-    private javax.swing.JTextField jTFCIRegistrarC;
     private javax.swing.JTextField jTFDirNegocio;
-    private javax.swing.JTextField jTFDireccionR;
     private javax.swing.JTextField jTFEliminarUser;
-    private javax.swing.JTextField jTFIBuscadorAct1;
     private javax.swing.JTextField jTFIBuscadorAct2;
     private javax.swing.JTextField jTFIBuscadorAct5;
     private javax.swing.JTextField jTFIBuscarRUC;
     private javax.swing.JTextField jTFIDemploye;
     private javax.swing.JTextField jTFNombreNegocio;
     private javax.swing.JTextField jTFNombreUser;
-    private javax.swing.JTextField jTFNombresR;
     private javax.swing.JTextField jTFRUCNegocio;
     private javax.swing.JTextField jTFSubtotal;
     private javax.swing.JTextField jTFTelefonoNegocio;
     private javax.swing.JTextField jTFTelefonoProvedor;
-    private javax.swing.JTextField jTFTelefonoR;
     private javax.swing.JTextField jTFnombreUser;
     private javax.swing.JTextField jTFnumerofactura;
-    private javax.swing.JTable jTIncidentes;
     private javax.swing.JTextField jTNombreCliente;
     private javax.swing.JTextField jTNombreDespachador;
     private javax.swing.JTextField jTNombreDespachadorAct;
@@ -8125,7 +6629,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JTextField jTNombreProducto1;
     private javax.swing.JTabbedPane jTPAdmin;
     private javax.swing.JTabbedPane jTPEmpleados;
-    private javax.swing.JTabbedPane jTPEmpleados1;
     private javax.swing.JTextField jTRUC;
     private javax.swing.JTextField jTRUCAct;
     private javax.swing.JTextField jTTelefonoCliente;
@@ -8134,7 +6637,6 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JTable jTablaAdmin1;
     private javax.swing.JTable jTablaAdmin2;
     private javax.swing.JTable jTablaAdmin3;
-    private javax.swing.JTable jTablaClientesR;
     private javax.swing.JTable jTablaProveedoresA;
     private javax.swing.JTable jTablaProveedoresCE;
     private javax.swing.JTable jTablaProveedoresR;
@@ -8155,13 +6657,8 @@ public class JFMenu extends javax.swing.JFrame {
     private javax.swing.JLabel menuLogout;
     private javax.swing.JLabel menuProveedores;
     private javax.swing.JLabel menuinventario;
-    private javax.swing.JTextField nombreActualizar;
     private javax.swing.JTextField nombreProducto;
-    private javax.swing.JCheckBox nombreSelect;
     private javax.swing.JPanel panelContent;
-    private javax.swing.JTable tablaActualizarCliente;
-    private javax.swing.JTextField telefonoActualizar;
-    private javax.swing.JCheckBox telefonoSelect;
     private javax.swing.JLabel txtDateLog;
     private javax.swing.JLabel txtID;
     // End of variables declaration//GEN-END:variables
